Refactor `parse_csv_option` to a dedicated utilities module.
Extracted `parse_csv_option` into a new `utilities/parse_csv.py` module for better reusability and modularity. Updated imports in affected files to reflect this change. This improves code organization and reduces duplication.
