+++ scm_config_clone/commands/objects/address.py
@@ -16,28 +16,7 @@ from scm.exceptions import (
 from scm.models.objects.address import AddressCreateModel, AddressResponseModel
 from tabulate import tabulate
 
-from scm_config_clone.utilities.settings import load_settings
-
-
-def parse_csv_option(value: Optional[str]) -> Optional[List[str]]:
-    """
-    Parse a comma-separated string into a list of stripped strings.
-
-    This utility function converts an option like "val1,val2,val3"
-    into ["val1", "val2", "val3"]. If the input string is None
-    or empty, it returns None.
-
-    Args:
-        value: The raw input string from a CLI option, possibly
-            containing comma-separated values.
-
-    Returns:
-        A list of strings if values are present, or None if the input
-        is empty or None.
-    """
-    if not value:
-        return None
-    return [v.strip() for v in value.split(",") if v.strip()]
+from scm_config_clone.utilities import load_settings, parse_csv_option
 
 
 def build_create_params(src_obj: AddressResponseModel, folder: str) -> Dict[str, Any]:
